feat(angular): produce build th Vite
