+++ packages/angular-table/tsconfig.json
@@ -12,5 +12,11 @@
     "experimentalDecorators": true,
     "importHelpers": true
   },
-  "include": ["src", "tests", "eslint.config.js", "vite.config.ts"]
+  "include": [
+    "src",
+    "tests",
+    "eslint.config.js",
+    "vite.config.ts",
+    "vitest.config.ts"
+  ]
 }
